tg3: Fix the setting of the APE_HAS_NCSI flag
Patchwork-id: 52771
O-Subject: [RHEL6.4 PATCH BZ 826718 47/55] tg3: Fix the setting of the APE_HAS_NCSI flag
Bugzilla: 826718
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Myron Stowe <mstowe@redhat.com>

http://bugzilla.redhat.com/show_bug.cgi?id=826718

commit 165f4d1cb3d7a10c5bc8c0fdae63323354f5a225
Author: Michael Chan <mchan@broadcom.com>
Date:   Mon, 16 Jul 2012 16:23:59 +0000 (+0000)

tg3: Fix the setting of the APE_HAS_NCSI flag

The driver currently skips setting this flag if the VPD contains the
firmware version string.  We fix this by separating the probing of NCSI
from the reading of the NCSI version string.  The APE_HAS_NCSI flag is
needed to properly read sensor data.

Signed-off-by: Michael Chan <mchan@broadcom.com>
Signed-off-by: David S. Miller <davem@davemloft.net>