mm: Continue to update file time in fault paths for OOT filesystems
Patchwork-id: 52848
O-Subject: [RHEL6.4 PATCH 06/34 V2] mm: Continue to update file time in fault paths for OOT filesystems
Bugzilla: 768304
RH-Acked-by: Dave Chinner <dchinner@redhat.com>

Resolves: BZ 768304 - Deadlock occurs under a race with "flush" kthread and ioctl() to thaw ext3, ext4, and xfs filesystems.

This is not an upstream patch, but contains RHEL6-specific code for
KABI preservation.

Add a new flag and helper to indicate that this filesystem will modify
time in its page_mkwrite.  (This flag is re-used to indicate the "new freeze"
mechanism later as well).  If not set, continue to do it in
the fault paths themselves, to support out of tree filesystems using the KABI
which expect fault paths to update the time.

The same is true of any vma which uses the generic filemap_page_mkwrite().

Flags will be set on our freeze-capable filesystems in later patches.

For review, it may be helpful to combine this patch with the previous to
see the net change.