openvswitch: Make Open vSwitch use non-accel VLAN tagging
Patchwork-id: 52852
O-Subject: [RHEL6.4 PATCH 2/2] openvswitch: Make Open vSwitch use non-accel VLAN tagging
Bugzilla: 859048
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Amerigo Wang <amwang@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

The upstream version of Open vSwitch is based on the VLAN hardware
acceleration rework of commit 3701e51 (vlan: Centralize handling of
hardware acceleration.) This commit allows Open vSwitch to call
__vlan_hwaccel_put_tag() and make use of acceleration if available.
The stack will automatically push the tag as vlan header in software
if acceleration is not available.

We do not want to backport this large infrastructure change in RHEL6
and therefore modify Open vSwitch to always use __vlan_put_tag() and
thus force software mode. Due to this we also don't have to convert
any vlan headers to vlan_tci as done by pop_vlan().

Why can't we use vlan_put_tag()? Because we don't know the outgoing
interface at the point of tagging. The tagging is performed before
the output action is executed.

This is a RHEL6 specific change as upstream is based on a different
VLAN infrastructure.

Signed-off-by: Thomas Graf <tgraf@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>