cgroups: don't attach task to subsystem if migration failed
Patchwork-id: 52892
O-Subject: [PATCH RHEL6.4 14/14] cgroups: don't attach task to subsystem if migration failed
Bugzilla: 751817
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>

Bugzilla: 751817
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=4899784

Upstream: 77ceab8ea590d7dc6c8f055ce43dfebd74428107
Author: Ben Blum <bblum@andrew.cmu.edu>
cgroups: don't attach task to subsystem if migration failed

If a task has exited to the point it has called cgroup_exit() already,
then we can't migrate it to another cgroup anymore.

This can happen when we are attaching a task to a new cgroup between the
call to ->can_attach_task() on subsystems and the migration that is
eventually tried in cgroup_task_migrate().

In this case cgroup_task_migrate() returns -ESRCH and we don't want to
attach the task to the subsystems because the attachment to the new cgroup
itself failed.

Fix this by only calling ->attach_task() on the subsystems if the cgroup
migration succeeded.

Reported-by: Oleg Nesterov <oleg@redhat.com>
Signed-off-by: Ben Blum <bblum@andrew.cmu.edu>
Acked-by: Paul Menage <paul@paulmenage.org>
Cc: Li Zefan <lizf@cn.fujitsu.com>
Cc: Tejun Heo <tj@kernel.org>
Signed-off-by: Frederic Weisbecker <fweisbec@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Frederic Weisbecker <fweisbec@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>