ext4: Add flags and structure used in ext4_ext_map_blocks()
Patchwork-id: 52910
O-Subject: [RHEL6.4 v2 PATCH 09/41] ext4: Add flags and structure used in ext4_ext_map_blocks()
Bugzilla: 739101
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>

BZ 739101
https://bugzilla.redhat.com/show_bug.cgi?id=739101

BREW 4887571
https://brewweb.devel.redhat.com/taskinfo?taskID=4887571

This commit backports on a fraction of upstream commit
e35fd6609b2fee54484d520deccb8f18bf7d38f3. In fact we backport only flags
and ext4_map_blocks structure, because we need this to use with the
punch hole code, but on the other hand we do not want to introduce all
the changes associated with this upstream commit.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>