ext4: fix hole punch failure when depth is greater than 0
Patchwork-id: 52921
O-Subject: [RHEL6.4 v2 PATCH 38/41] ext4: fix hole punch failure when depth is greater than 0
Bugzilla: 739101
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

From: Ashish Sangwan <ashishsangwan2@gmail.com>

BZ 739101
https://bugzilla.redhat.com/show_bug.cgi?id=739101

BREW 4887571
https://brewweb.devel.redhat.com/taskinfo?taskID=4887571

Upstream commit 968dee77220768a5f52cf8b21d0bdb73486febef

Whether to continue removing extents or not is decided by the return
value of function ext4_ext_more_to_rm() which checks 2 conditions:
a) if there are no more indexes to process.
b) if the number of entries are decreased in the header of "depth -1".

In case of hole punch, if the last block to be removed is not part of
the last extent index than this index will not be deleted, hence the
number of valid entries in the extent header of "depth - 1" will
remain as it is and ext4_ext_more_to_rm will return 0 although the
required blocks are not yet removed.

This patch fixes the above mentioned problem as instead of removing
the extents from the end of file, it starts removing the blocks from
the particular extent from which removing blocks is actually required
and continue backward until done.

Signed-off-by: Ashish Sangwan <ashish.sangwan2@gmail.com>
Signed-off-by: Namjae Jeon <linkinjeon@gmail.com>
Reviewed-by: Lukas Czerner <lczerner@redhat.com>
Cc: stable@vger.kernel.org
Signed-off-by: Jarod Wilson <jarod@redhat.com>