ext4: remove restrictive checks for EOFBLOCKS_FL
Patchwork-id: 52935
O-Subject: [RHEL6.4 v2 PATCH 40/41] ext4: remove restrictive checks for EOFBLOCKS_FL
Bugzilla: 739101
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

BZ 739101
https://bugzilla.redhat.com/show_bug.cgi?id=739101

BREW 4887571
https://brewweb.devel.redhat.com/taskinfo?taskID=4887571

Upstream commit afcff5d80a4106e732d903640161d23950eb8e3b

We are going to remove the EOFBLOCKS_FL flag in the future, so this is
the first part of the removal. We can not remove it entirely just now,
since the e2fsck is still checking for it and it might cause headache to
some people. Instead, remove the restrictive checks now and the rest
later, when the new e2fsck code is out and common enough.

This is also needed because punch hole already breaks the EOFBLOCKS_FL
semantics, so it might cause the some troubles. So simply remove it.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Jarod Wilson <jarod@redhat.com>