hpwdt: Unregister NMI events on exit
Patchwork-id: 52956
O-Subject: [RHEL6 PATCH 3/4] watchdog: hpwdt: Unregister NMI events on exit.
Bugzilla: 818372
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=818372

commit a089361cf5f1d6a5295aa5385238bd044998e1e9
Author: Mingarelli, Thomas <Thomas.Mingarelli@hp.com>
Date:   Tue Jun 26 10:27:00 2012 +0200

    watchdog: hpwdt: Unregister NMI events on exit.

    This patch is to unregister for NMI events upon exit. Also we are now
    making the default setting for allow_kdump enabled.

    Signed-off-by: Thomas Mingarelli <thomas.mingarelli@hp.com>
    Signed-off-by: Wim Van Sebroeck <wim@iguana.be>

The only interesting piece of this upstream commit is the enabling of the
allow_kdump flag.  The other piece is not applicable to RHEL-6.

Setting allow_kdump to 1 is a small behaviour change, but as it only disables
the external hpwdt timer when panic'ing into kdump, it seemed reasonable. And
considering, we try to make sure all kernel panics lead to kdump, it seemed
like a smarter change.

Signed-off-by: Jarod Wilson <jarod@redhat.com>