sfc: Remove configurable FIFO thresholds for pause frame generation
Patchwork-id: 52990
O-Subject: [RHEL6.4 PATCH BZ 824042 011/142] sfc: Remove configurable FIFO thresholds for pause frame generation
Bugzilla: 824042
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>

From: Ben Hutchings <bhutchings@solarflare.com>

In Falcon we can configure the fill levels of the RX data FIFO which
trigger the generation of pause frames (if enabled), and we have
module parameters for this.

Siena does not allow the levels to be configured (or, if it does, this
is done by the MC firmware and is not configurable by drivers).

So far as I can tell, the module parameters are not used by our
internal scripts and have not been documented (with the exception of
the short parameter descriptions).  Therefore, remove them and always
initialise Falcon with the default values.

Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>
(cherry picked from commit 5fb6b06d4eda2167eab662ad5e30058cecd67b8b)

Signed-off-by: Jarod Wilson <jarod@redhat.com>