Revert: intel_idle: disable NHM/WSM HW C-state auto-demotion
Patchwork-id: 53455
O-Subject: [RHEL 6.4 PATCH] Revert "[x86] intel_idle: disable NHM/WSM HW C-state auto-demotion"
Bugzilla: 834097
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

rhbz: #834097

This reverts commit 134b1492bebe682a0b4f7042a70cccc732dda5de. In theory
cpuidle is smart enough to detect patterns and avoid poor choice of C
states - in practice it seems that the CPU is doing a better job of it at
the moment. Re-enable autodemotion in order to avoid the performance
regression introduced by this patch in the first place.

Signed-off-by: Matthew Garrett <mjg@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>