perf/symbols: Follow .gnu_debuglink section to find separate symbols
Patchwork-id: 53752
O-Subject: [PATCH RHEL6.4 BZ860404 26/74] perf symbols: Follow .gnu_debuglink section to find separate symbols
Bugzilla: 860404
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 860404
https://bugzilla.redhat.com/show_bug.cgi?id=860404

upstream status
===============
commit 209bd9e3e14712d74c8bebb028afda905d689f1c
Author: Pierre-Loup A. Griffais <pgriffais@nvidia.com>
Date:   Fri Jun 22 11:38:13 2012 -0700

description
===========
The .gnu_debuglink section is specified to contain the filename of the
debug info file, as well as a CRC that can be used to validate it.

This doesn't currently use the checksum and relies on the usual build-id
matching for validation.

This provides more context:
http://sourceware.org/gdb/onlinedocs/gdb/Separate-Debug-Files.html

Signed-off-by: Jarod Wilson <jarod@redhat.com>