perf/report: Delay sample_type checks in pipe mode
Patchwork-id: 53748
O-Subject: [PATCH RHEL6.4 BZ860404 28/74] perf report: Delay sample_type checks in pipe mode
Bugzilla: 860404
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 860404
https://bugzilla.redhat.com/show_bug.cgi?id=860404

upstream status
===============
commit 300aa941650e98966ad85847527537df5b11a87e
Author: David Ahern <dsahern@gmail.com>
Date:   Mon Jun 11 13:48:41 2012 -0600

description
===========
The pipeline:
  perf record -a -g -o - sleep 5 |perf inject -v -b  | perf report  -g -i -

generates the warning:
  Selected -g but no callchain data. Did you call 'perf record' without -g?

The problem is that the header data is not written to the pipe, so the
sample_type has not been available when perf_report__setup_sample_type
is called. For pipe mode, record dumps the sample type as part of the
synthesized events stream -- perf_event__synthesize_attrs(). Handle this
be detecting pipe mode and not doing early sanity checks on sample_type.

Signed-off-by: Jarod Wilson <jarod@redhat.com>