perf/kvm: Set name for VM process in guest machine
Patchwork-id: 53756
O-Subject: [PATCH RHEL6.4 BZ860404 40/74] perf kvm: Set name for VM process in guest machine
Bugzilla: 860404
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 860404
https://bugzilla.redhat.com/show_bug.cgi?id=860404

upstream status
===============
commit 5cd95c2db479aa7a66f6fa572dfa410c6314c78e
Author: David Ahern <dsahern@gmail.com>
Date:   Fri Jul 20 17:25:47 2012 -0600

description
===========
COMM events are not generated in the context of a guest machine, so the
thread name is never set for the VMM process. For example, the qemu-kvm
name applies to the process in the host machine, not the guest machine.
So, samples for guest machines are currently displayed as:

    99.67%     :5671  [unknown]         [g] 0xffffffff81366b41

where 5671 is the pid of the VMM. With this patch the samples in the guest
machine are shown as:

    18.43%  [guest/5671]  [unknown]           [g] 0xffffffff810d68b7

Signed-off-by: Jarod Wilson <jarod@redhat.com>