perf/hists: Return correct number of characters printed in callchain
Patchwork-id: 53788
O-Subject: [PATCH RHEL6.4 BZ860404 45/74] perf hists: Return correct number of characters printed in callchain
Bugzilla: 860404
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 860404
https://bugzilla.redhat.com/show_bug.cgi?id=860404

upstream status
===============
commit 8760db726e2afcd1a78e2ff58965c6b35a5826cb
Author: Frederic Weisbecker <fweisbec@gmail.com>
Date:   Wed Jul 18 19:10:55 2012 +0200

description
===========
Include the omitted number of characters printed for the first entry.

Not that it really matters because nobody seem to care about the number
of printed characters for now. But just in case.

Signed-off-by: Jarod Wilson <jarod@redhat.com>