perf/kvm: Use strtol for walking guestmount directory
Patchwork-id: 53800
O-Subject: [PATCH RHEL6.4 BZ860404 68/74] perf kvm: Use strtol for walking guestmount directory
Bugzilla: 860404
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 860404
https://bugzilla.redhat.com/show_bug.cgi?id=860404

upstream status
===============
commit 347ed9903a10179d1cf733d5d77072b283d89da3
Author: David Ahern <dsahern@gmail.com>
Date:   Sun Jul 29 20:54:35 2012 -0600

description
===========
Only want to process directories under the guestmnount directory that
have a pid as a name (ie, all digits). Other entries in the guestmount
directory should be ignored.  There is already a check that requires the
first character of each entry to be a digit, but atoi is used to convert
the directory name to a pid. For example if guestmount contains a
directory with the name 1foo, atoi converts it to a pid of 1 and a
machine is created with a pid of 1. This is wrong; this directory really
should be ignored. Use strtol to do that.

Signed-off-by: Jarod Wilson <jarod@redhat.com>