cpu: keep SEP flag enabled on 64-bit if execshield is on
Patchwork-id: 53836
O-Subject: [RHEL6.4 PATCH] keep SEP flag enabled on 64-bit if execshield is on
Bugzilla: 821463
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Gleb Natapov <gleb@redhat.com>

Bugzilla: 821463

When execshield is enabled (the default on RHEL-6), the SEP feature bit
is always being cleared by the kernel on 64-bit mode, but it should be
cleared only if the NX feature is unavailable.

This affects KVM, making it not expose the SEP feature to any guest.

The desired behavior is:

                     |   keep SEP bit?
arch | XS | PAE | NX | current | desired
-----+----+-----+----+---------+--------
 32  |  0 |   - |  - | 1       | 1
 32  |  1 |   0 |  - | 0       | 0
 32  |  1 |   1 |  0 | 0       | 0
 32  |  1 |   1 |  1 | 1       | 1
 64  |  0 |   - |  - | 1       | 1
 64  |  1 |   0 |  0 | 0       | 0
 64  |  1 |   0 |  1 | 0       | 1 (here is the current bug)

Legend:
  XS:  exec_shield != 0
  PAE: defined(CONFIG_X86_PAE)
  NX:  test_cpu_cap(c, X86_FEATURE_NX

This patch changes the code to match the desired behavior.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Don Zickus <dzickus@redhat.com>
Cc: Larry Woodman <lwoodman@redhat.com>
Cc: Rik van Riel <riel@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>