loop: add discard support for loop devices
Patchwork-id: 53841
O-Subject: [RHEL6.4 PATCH 1/4 v2] loop: add discard support for loop devices
Bugzilla: 766815
RH-Acked-by: Dave Chinner <dchinner@redhat.com>
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

BZ 766815
https://bugzilla.redhat.com/show_bug.cgi?id=766815

BREW 4887571
https://brewweb.devel.redhat.com/taskinfo?taskID=4887571

This is backport of upstream commit
dfaa2ef68e80c378e610e3c8c536f1c239e8d3ef. Small change is required
because rhel kernel still register ->fallocate as inode operation
whereas upstream switched to use file operation. This however can not be
done easily in RHEL because of kABI.

This commit adds discard support for loop devices. Discard is usually
supported by SSD and thinly provisioned devices as a method for
reclaiming unused space. This is no different than trying to reclaim
back space which is not used by the file system on the image, but it
still occupies space on the host file system.

We can do the reclamation on file system which does support hole
punching. So when discard request gets to the loop driver we can
translate that to punch a hole to the underlying file, hence reclaim
the free space.

This is very useful for trimming down the size of the image to only what
is really used by the file system on that image. Fstrim may be used for
that purpose.

It has been tested on ext4, xfs and btrfs with the image file systems
ext4, ext3, xfs and btrfs. ext4, or ext6 image on ext4 file system has
some problems but it seems that ext4 punch hole implementation is
somewhat flawed and it is unrelated to this commit.

Also this is a very good method of validating file systems punch hole
implementation.

Note that when encryption is used, discard support is disabled, because
using it might leak some information useful for possible attacker.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Reviewed-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Jens Axboe <jaxboe@fusionio.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>