migrate: fix kABI issues
Patchwork-id: 53885
O-Subject: [RHEL6.4 PATCH 21/21] fix kABI issues (RHEL only)
Bugzilla: 713546 783248
RH-Acked-by: Jerome Marchand <jmarchan@redhat.com>
RH-Acked-by: Johannes Weiner <jweiner@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

The ->migratepage operation gained an additional argument earlier in
the series. This breaks the kABI. The obvious fix is to remove that
argument.

However, we still want to pass information down on whether we want
to do a synchronous or asynchronous migration.

The workaround is to overload PF_SWAPWRITE to indicate whether we
are doing a synchronous migration. It does not look like the two
callchains intersect. However, even if they would the result is at
most a reduction in the performance gain from this patch series.

Fixes bugs 783248 & 713546

Signed-off-by: Rik van Riel <riel@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>