spec: latest bfa driver requires newer firmware
Patchwork-id: 53977
O-Subject: [RHEL6 PATCH] redhat: latest bfa driver requires newer firmware
Bugzilla: 861422
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Kyle McMartin <kmcmarti@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=861422

The bfa driver update in kernel-2.6.32-308.el6 fails to install on systems
with the rhel6.3 bfa-firmware. There's an updated bfa-firmware package
available that remedy's this, but to ensure people don't install and
reboot with the older firmware in place, we need to add a Conflicts marker
to the kernel rpm, to trigger bfa-firmware also upgrading when a newer
kernel that needs it is installed.