be2net: Regression bug wherein VFs creation broken for multiple cards
Patchwork-id: 54018
O-Subject: [RHEL6.4 PATCH v2 3/6] be2net: Regression bug wherein VFs creation broken for multiple cards.
Bugzilla: 819782
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

BZ: #819782

Backport note:
The upstream commit tries to fix a bug in VFs enumeration, but this fix was not
complete and was replaced by proper recent fix (last in this serie). This commit
is present in this serie for completness.

Upstream commit:
commit 7665de15642dfb1709177517aa0488162727342c
Author: Somnath Kotur <somnath.kotur@emulex.com>
Date:   Sun Jun 24 19:42:00 2012 +0000

    be2net: Regression bug wherein VFs creation broken for multiple cards.

    Fix be_find_vfs() to check for matching bus number as well along with devfn

    Signed-off-by: Somnath Kotur <somnath.kotur@emulex.com>
    Signed-off-by: David S. Miller <davem@davemloft.net>

Signed-off-by: Ivan Vecera <ivecera@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>