be2net: fix max VFs reported by HW
Patchwork-id: 54015
O-Subject: [RHEL6.4 PATCH v2 4/6] be2net: fix max VFs reported by HW
Bugzilla: 819782
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

BZ: #819782

Backport note:
This commit is not much related to the BZ but it's more or less dependent on
the 1st patch in this serie. So I have included it in this set instead of
driver update set.

Upstream commit:
commit 7c5a524221dc3cf599a6aa4773eb0c1e2ca38adb
Author: Vasundhara Volam <vasundhara.volam@emulex.com>
Date:   Tue Aug 28 20:37:41 2012 +0000

    be2net: fix max VFs reported by HW

    BE3 FW allocates VF resources for upto 30 VFs per PF while a max value of 32
    may be reported via PCI config space. Fix this in the driver.

    Signed-off-by: Vasundhara Volam <vasundhara.volam@emulex.com>
    Signed-off-by: Sathya Perla <sathya.perla@emulex.com>
    Signed-off-by: David S. Miller <davem@davemloft.net>

Signed-off-by: Ivan Vecera <ivecera@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>